refactor: extract plugins-compat-routes.ts from server.ts
Move 4 plugin endpoints (GET/PUT /api/plugins, test, reveal) + entire
helper cluster (normalizePluginId, buildPluginListResponse,
persistCompatPluginMutation, resolvePluginManifestPath, etc.) to
plugins-compat-routes.ts.

server.ts: 3,974 → 3,238 lines (-736)
plugins-compat-routes.ts: 804 lines (new)
Cumulative server.ts reduction: 4,269 → 3,238 (-24.2%)
